requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
nltk==3.5
regex==2020.11.13
requests==2.24.0
soupsieve==2.0.1
tqdm==4.53.0
urllib3==1.25.11
Werkzeug==1.0.1
certifi==2020.12.5
chardet==4.0.0
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.1
decorator==4.4.2
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
future==0.18.2
httplib2==0.19.0
idna==2.10
isodate==0.6.0
lxml==4.6.2
networkx==2.5
nibabel==3.2.1
nipype==1.6.0
numpy==1.20.1
packaging==20.9
pandas==1.1.5
pathlib==1.0.1
pdf2image==1.14.0
Pillow==8.1.0
prov==2.0.0
pydot==1.4.1
PyMuPDF==1.18.8
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
pyxnat==1.4
rdflib==5.0.0
requests==2.25.1
scipy==1.6.0
simplejson==3.17.2
six==1.15.0
traits==6.2.0
urllib3==1.26.3
